Added updatedBp method and did some restrucuring
