+++ backend/routes/bp.go
@@ -5,10 +5,11 @@ import (
 	"github.com/katamyra/kubestellarUI/wds/bp"
 )
 
-func setupBlueprintRoutes(router *gin.Engine) {
+func setupBindingPolicyRoutes(router *gin.Engine) {
 	router.GET("/api/bp", bp.GetAllBp)
 	router.GET("/api/bp/status", bp.GetBpStatus)
 	router.POST("/api/bp/create", bp.CreateBp)
 	router.DELETE("/api/bp/delete/:name", bp.DeleteBp)
 	router.DELETE("/api/bp/delete", bp.DeleteAllBp)
+	router.PATCH("/api/bp/patch", bp.UpdateBp)
 }
