+++ backend/routes/setup.go
@@ -10,6 +10,6 @@ func SetupRoutes(router *gin.Engine) {
 	setupDeploymentRoutes(router)
 	setupServiceRoutes(router)
 	setupNamespaceRoutes(router)
-	setupBlueprintRoutes(router)
 	SetupAuthRoutes(router)
+	setupBindingPolicyRoutes(router)
 }
